Moved some editor functionality to separate file
